# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "ASP - v2",
    "contact": {}
  },
  "host": "example.com",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/organizations": {
      "post": {
        "summary": "Create Organization",
        "tags": [
          "Organization"
        ],
        "operationId": "CreateOrganization",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateOrganizationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "get": {
        "summary": "Get Organizations",
        "tags": [
          "Organization"
        ],
        "operationId": "GetOrganizations",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/systems": {
      "post": {
        "summary": "Create System",
        "tags": [
          "Organization"
        ],
        "operationId": "CreateSystem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreateSystemRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "Get Systems",
        "tags": [
          "Organization"
        ],
        "operationId": "GetSystems",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/systems/2/environments": {
      "post": {
        "summary": "Creae Environment",
        "tags": [
          "Organization"
        ],
        "operationId": "CreaeEnvironment",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/CreaeEnvironmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "Get Environments",
        "tags": [
          "Organization"
        ],
        "operationId": "GetEnvironments",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
...
    "CreateOrganizationRequest": {
      "title": "CreateOrganizationRequest",
      "example": {
        "name": "{{$randomCompanyName}}"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
    "CreateSystemRequest": {
      "title": "CreateSystemRequest",
      "example": {
        "name": "{{$randomBankAccountName}}"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
    "CreaeEnvironmentRequest": {
      "title": "CreaeEnvironmentRequest",
      "example": {
        "name": "Develop"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
...
  },
...
}